Refactoring

+++ db/migrate/20200414035616_create_match_statistics.rb
@@ -3,8 +3,8 @@ class CreateMatchStatistics < ActiveRecord::Migration[5.2]
     create_table :match_statistics do |t|
       t.references :player, foreign_key: true
       t.references :match, foreign_key: true
-      t.decimal :distance_covered, null: false, default: 0
-      t.decimal :productivity, null: false, default: 0
+      t.float :distance_covered, null: false, default: 0
+      t.float :productivity, null: false, default: 0
 
       t.timestamps
     end
